+++ plugins/source/k8s/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [6.1.0](https://github.com/cloudquery/cloudquery/compare/plugins-source-k8s-v6.0.17...plugins-source-k8s-v6.1.0) (2024-06-11)
+
+
+### Features
+
+* Remove logging of error events to Sentry in plugins ([#18165](https://github.com/cloudquery/cloudquery/issues/18165)) ([fc4ff27](https://github.com/cloudquery/cloudquery/commit/fc4ff27d37f9250b4cf912474073169406cb01fa))
+
 ## [6.0.17](https://github.com/cloudquery/cloudquery/compare/plugins-source-k8s-v6.0.16...plugins-source-k8s-v6.0.17) (2024-06-04)